requirements.txt
pymupdf
sentence-transformers
chromadb
google-generativeai
rank-bm25
python-dotenv
